Make constructor mapping a little more lenient
Before, all attempts at mapping constructor parameters for classes that
did not have a constructor defined, or where no defined constructor had
a parameter with that name, would throw an exception with a pretty poor
uninformative exception message.

These scenarios are now considered valid mapping configurations, as long
as the stuff that's needed to actually perform a mapping is also
configured.

Fixes #1528